randomop - added isMinMaxSet boolean and fixed a few other things
Turns out that the new boolean is unnecessary, so I will have to
revisit this. Added another overloaded version of setMinMax.

Fixed a bug in findMatchPair, where we need to use the size of
the m_qlPrRepeats list instead of the loop index in a conditional.

Added logic to getOneUnique to limit ones and zeros.

Fixed a bug in findMatch() where m_maxNumRightOps was being used
where m_maxNumOperandPairs should be used.

This stuff is still not quite there, yet. It may need some
rethinking and maybe even a bit of redesign.